feat: split eos
