+++ doc.md
@@ -5,12 +5,6 @@
 - [ERC20](./src/eth/erc20.ts)
 - [Parity](./src/eth/parity.ts)
 - [OmniLayer](./src/usdt/client.ts)
-- [EOS](./src/eos/rpc.ts)
-
-### EOS Actions
-
-- [eosio.token](https://github.com/isLishude/eos-actions-types/blob/master/eosio.token.d.ts)
-- [eosio](https://github.com/isLishude/eos-actions-types/blob/master/eosio.d.ts)
 
 ## Example
