+++ src/omni/type.ts
