--- test/eos/client_test.ts
@@ -1,5 +1,0 @@
-import test from "ava";
-
-test("test it", (t) => {
-  t.pass();
-});
